src/Server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
click==8.1.7
ecdsa==0.19.0
fastapi==0.110.2
h11==0.14.0
httptools==0.6.1
idna==3.7
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
